==> Projet_BlindCV/Part1_BlindCV.ipynb <==
# PART 1 - BlindCV

### 1. Chargement des librairies et du dataset

import pandas as pd
import spacy

dataset = pd.read_csv("dataset_CV.csv", delimiter = ";")

dataset.head()

dataset.shape

dataset.tail()

## 2. Installation de Spacy

## Instanciation de la bibliothèque fr_core_news_lg de Spacy

nlp = spacy.load("fr_core_news_lg")

## 3. Pre-processing

dataset.head()

a. Création d'une nouvelle colonne et remplacement du "\n" dans les textes par un espace

dataset["Tokenization_preparation"] = dataset["CV_text"].apply(lambda x : str(x).replace("\n", " "))

dataset.head()

b. Tokenisation du texte

dataset["Tokenization"] = dataset["Tokenization_preparation"].apply(lambda x : nlp(x))

dataset.head()

c. Compter le nombre de tokens par CV

dataset["Number_of_tokens"] = dataset["Tokenization"].apply(lambda x : len(x))

dataset.head()

dataset.describe()

c. Obtenir la grammaire des mots du texte du CV_1

## Test sur 1 CV
dataset.iloc[0,2]

## Création d'une variable pour accueillir le texte du CV
doc = dataset.iloc[0,2]

## Application de la commande nlp pour tokeniser le texte du CV
doc = nlp(doc)

doc 

## Fin Part 1
