==> Projet_BlindCV/Part4_BlindCV_Labelise_features2_20.ipynb <==
## Part 10 - Création de features supplémentaires test

import pandas as pd
import spacy

# Chargement de la dataset
dataset = pd.read_csv("dataset_CV_labelise_features1_200.csv", delimiter = ";", encoding = "utf-8")
len(dataset)

dataset_sample = dataset.head(20)

dataset_sample.head()

nlp = spacy.load("fr_core_news_lg")

dataset_sample["Tokenization"] = dataset_sample["Sentences_CV_clean"].apply(lambda x : nlp(x))

def count_verb(text, nlp):
    increment  = 0
    tokens = nlp(text)
    for token in tokens:
        if token.pos_ == 'VERB':
            increment = increment + 1
    return increment

def count_noun(text, nlp):
    increment  = 0
    tokens = nlp(text)
    for token in tokens:
        if token.pos_ == 'NOUN':
            increment = increment + 1
    return increment

def count_propn(text, nlp):
    increment  = 0
    tokens = nlp(text)
    for token in tokens:
        if token.pos_ == 'PROPN':
            increment = increment + 1
    return increment

def count_num(text, nlp):
    increment  = 0
    tokens = nlp(text)
    for token in tokens:
        if token.pos_ == 'NUM':
            increment = increment + 1
    return increment

dataset_sample["Verb_count"] = dataset_sample["Tokenization"].apply(lambda x : count_verb(x, nlp))
dataset_sample["Propn_count"] = dataset_sample["Tokenization"].apply(lambda x : count_propn(x, nlp))
dataset_sample["Noun_count"] = dataset_sample["Tokenization"].apply(lambda x : count_noun(x, nlp))
dataset_sample["Num_count"] = dataset_sample["Tokenization"].apply(lambda x : count_num(x, nlp))

dataset_sample["Pourcentage_verb_sentence"] = dataset_sample["Verb_count"] / dataset_sample["Nb_tokens"]
dataset_sample["Pourcentage_propn_sentence"] = dataset_sample["Propn_count"] / dataset_sample["Nb_tokens"]
dataset_sample["Pourcentage_noun_sentence"] = dataset_sample["Noun_count"] / dataset_sample["Nb_tokens"]
dataset_sample["Pourcentage_num_sentence"] = dataset_sample["Num_count"] / dataset_sample["Nb_tokens"]

dataset_sample.head()

## Modification de l'index de la colonne Label pour la positionner en derniere colonne

dataset_sample["Label2"] = dataset_sample["Label"]
del dataset_sample["Label"]

dataset_sample["Label"] = dataset_sample["Label2"]
del dataset_sample["Label2"]
