==> Projet_BlindCV/Part5_BlindCV_Labelise_features2_200.ipynb <==
## Part 10 - Création de features supplémentaires test

import pandas as pd
import spacy

# Chargement de la dataset
dataset = pd.read_csv("dataset_CV_labelise_features1_200.csv", delimiter = ";", encoding = "utf-8")
len(dataset)

dataset.head()

nlp = spacy.load("fr_core_news_lg")

dataset["Tokenization"] = dataset["Sentences_CV_clean"].astype(str).apply(lambda x : nlp(x))

def count_verb(text, nlp):
    increment  = 0
    tokens = nlp(text)
    for token in tokens:
        if token.pos_ == 'VERB':
            increment = increment + 1
    return increment

def count_noun(text, nlp):
    increment  = 0
    tokens = nlp(text)
    for token in tokens:
        if token.pos_ == 'NOUN':
            increment = increment + 1
    return increment

def count_propn(text, nlp):
    increment  = 0
    tokens = nlp(text)
    for token in tokens:
        if token.pos_ == 'PROPN':
            increment = increment + 1
    return increment

def count_num(text, nlp):
    increment  = 0
    tokens = nlp(text)
    for token in tokens:
        if token.pos_ == 'NUM':
            increment = increment + 1
    return increment

dataset["Verb_count"] = dataset["Tokenization"].apply(lambda x : count_verb(x, nlp))
dataset["Propn_count"] = dataset["Tokenization"].apply(lambda x : count_propn(x, nlp))
dataset["Noun_count"] = dataset["Tokenization"].apply(lambda x : count_noun(x, nlp))
dataset["Num_count"] = dataset["Tokenization"].apply(lambda x : count_num(x, nlp))

dataset["Pourcentage_verb_sentence"] = dataset["Verb_count"] / dataset["Nb_tokens"]
dataset["Pourcentage_propn_sentence"] = dataset["Propn_count"] / dataset["Nb_tokens"]
dataset["Pourcentage_noun_sentence"] = dataset["Noun_count"] / dataset["Nb_tokens"]
dataset["Pourcentage_num_sentence"] = dataset["Num_count"] / dataset["Nb_tokens"]

## Modification de l'index de la colonne Label pour la positionner en derniere colonne

dataset["Label2"] = dataset["Label"]
del dataset["Label"]

dataset["Label"] = dataset["Label2"]
del dataset["Label2"]

dataset.head()

dataset.shape

## Création d'un fichier Excel avec toutes les phrases de tous les CV
PATH = "/Users/kinn/Desktop/Projet_BlindCV/dataset_CV_labelise_features2_200.csv"
dataset.to_csv(path_or_buf = PATH, index = False, sep = ";")
